commit 0.56 Updated Task functionality

+++ src/main/java/org/example/kihelp_back/task/usecase/TaskDeleteUseCase.java
@@ -1,5 +1,5 @@
 package org.example.kihelp_back.task.usecase;
 
 public interface TaskDeleteUseCase {
-    void deleteTask(Integer taskId);
+    void deleteTask(Long taskId);
 }
